move errors to dedicated file
