+++ src/lexer.rs
@@ -1,6 +1,6 @@
-use std::fmt;
 use std::fmt::Debug;
 
+use crate::error::LexerError;
 use crate::scanner::{Scanner, ScannerState};
 
 #[derive(Debug, Clone, PartialEq)]
@@ -69,25 +69,6 @@ pub trait Tokenizer: Scanner {
     fn scan_token(&mut self) -> Result<(), Self::Error>;
     fn add_token(&mut self, token_type: Self::TokenType);
 }
-
-#[derive(Debug)]
-pub enum LexerError {
-    EmptyToken(usize),
-    UnexpectedCharacter(char, usize),
-}
-
-impl fmt::Display for LexerError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            LexerError::EmptyToken(line) => write!(f, "Empty token at line {}", line),
-            LexerError::UnexpectedCharacter(c, line) => {
-                write!(f, "Unexpected character '{}' at line {}", c, line)
-            }
-        }
-    }
-}
-
-impl std::error::Error for LexerError {}
 
 pub struct Lexer {
     source: String,
