+++ src/lib.rs
@@ -1,2 +1,3 @@
+mod error;
 mod lexer;
 mod scanner;
